requirements.txt
openai
python-dotenv
flask
flask-cors
openai
python-dotenv
